+++ tests/health_check.py
@@ -18,7 +18,7 @@ from orchard.core.metadata import DatasetRegistryWrapper
 from orchard.core.paths import HEALTHCHECK_LOGGER_NAME
 from orchard.data_handler.data_explorer import show_samples_for_dataset
 from orchard.data_handler.diagnostic import create_temp_loader
-from orchard.data_handler.fetcher import load_dataset_health_check
+from orchard.data_handler.dispatcher import load_dataset_health_check
 
 # Logging Setup
 logger = logging.getLogger(HEALTHCHECK_LOGGER_NAME)